# production-schema/prototype_openapi.schema.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "description": "\nTest server to store and retrieve log entries from volitile memory.\n\n## Log entries\n\nYou can:\n* **create log entries**.\n* **read log entries**.\n",
    "version": "0.0.2"
  },
  "paths": {
    "/logs/entries": {
      "get": {
        "tags": [
          "log entries"
        ],
        "summary": "Read Entries",
        "operationId": "read_entries_logs_entries_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Entry Ids",
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "name": "entry_ids",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Device Ids",
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "name": "device_ids",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Entries Logs Entries Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetLogEntryResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "log entries"
        ],
        "summary": "Add Entry",
        "operationId": "add_entry_logs_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostLogEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GetLogEntryResponse": {
        "title": "GetLogEntryResponse",
        "required": [
          "entry_id",
          "entry"
        ],
        "type": "object",
        "properties": {
          "entry_id": {
            "title": "Entry Id",
            "type": "integer"
          },
          "entry": {
            "$ref": "#/components/schemas/LogEntry"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Location": {
        "title": "Location",
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          }
        }
      },
      "LogEntry": {
        "title": "LogEntry",
        "required": [
          "device_id",
          "title",
          "entry"
        ],
        "type": "object",
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "entry": {
            "title": "Entry",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          }
        }
      },
      "PostLogEntryResponse": {
        "title": "PostLogEntryResponse",
        "required": [
          "entry_id"
        ],
        "type": "object",
        "properties": {
          "entry_id": {
            "title": "Entry Id",
            "type": "integer"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "log entries",
      "description": "Manage log entries."
    }
  ]
}